+++ packages/map-template/src/components/Map/Map.jsx
@@ -146,6 +146,10 @@ function Map({ onLocationClick, onVenueChangedOnMap, useMapProviderModule }) {
                 setVenue(venueToShow, mapsIndoorsInstance).then(() => {
                     onVenueChangedOnMap(venueToShow);
                 });
+            } else if (venueToShow && !isLocationClicked && !locationId && kioskOriginLocationId && !isDesktop) {
+                setVenue(venueToShow, mapsIndoorsInstance).then(() => {
+                    onVenueChangedOnMap(venueToShow);
+                });
             } else if (venueToShow) {
                 onVenueChangedOnMap(venueToShow);
             }
@@ -261,6 +265,8 @@ function Map({ onLocationClick, onVenueChangedOnMap, useMapProviderModule }) {
 
         const venueToShow = getVenueToShow(venueName, venues);
         if (venueToShow && !locationId && !kioskOriginLocationId) {
+            setVenue(venueToShow, miInstance);
+        } else if (venueToShow && !locationId && kioskOriginLocationId && !isDesktop) {
             setVenue(venueToShow, miInstance);
         }
     };